+++ pipeline/pipe/h/hipfile/environment.py
@@ -13,16 +13,8 @@ from pipe.environment.version_adapter import (
     path_matches_stream,
 )
 from pipe.glui.dialogs import MessageDialog
-from pipe.glui.save_version_dialog import PromoteVersionDialog, SaveVersionDialog
-from pipe.glui.version_browser import VersionBrowserWidget
 from pipe.struct.db import Environment, SGEntity, normalize_display_name
-from pipe.versioning import (
-    VersionStreamSpec,
-    list_version_records,
-    promote_version as promote_version_record,
-    save_version as save_version_record,
-    version_label,
-)
+from pipe.versioning import VersionStreamSpec
 
 from .filemanager import HFileManager
 
@@ -32,6 +24,9 @@ log = logging.getLogger(__name__)
 class HEnvFileManager(HFileManager):
     def __init__(self) -> None:
         super().__init__(Environment)
+
+    def _entity_label(self) -> str:
+        return "set"
 
     def _generate_filename_ext(self, entity) -> tuple[str, str]:
         env = cast(Environment, entity)
@@ -171,153 +166,25 @@ class HEnvFileManager(HFileManager):
             return None
         return environment, stream
 
-    def open_version_browser(self) -> None:
-        hip_path = self._current_hip_path()
-        if hip_path is None:
-            MessageDialog(
-                self._main_window,
-                "No valid set HIP is open. Use Open Set first.",
-                "Version History",
-            ).exec_()
-            return
-
+    def _resolve_current_stream(
+        self, hip_path: Path
+    ) -> tuple[VersionStreamSpec, str, SGEntity] | None:
         resolved = self._resolve_current_set_stream(hip_path)
         if resolved is None:
-            MessageDialog(
-                self._main_window,
-                "Could not resolve the current HIP to a valid set file. Use Open Set first.",
-                "Version History",
-            ).exec_()
-            return
-
-        environment, set_stream = resolved
-        records = list_version_records(set_stream)
-        if not records:
-            MessageDialog(
-                self._main_window,
-                "No version history was found for this set.",
-                "No Versions",
-            ).exec_()
-            return
-
-        browser = VersionBrowserWidget(
-            self._main_window,
-            records,
-            owner_label=environment.display_name or environment.name or "Set",
-            stream_label=set_stream.label,
+            return None
+        environment, stream = resolved
+        return (
+            stream,
+            environment.display_name or environment.name or "Set",
+            environment,
         )
-        if not browser.exec_():
-            return
-
-        selected_record = browser.get_selected_record()
-        selected_action = browser.get_selected_action()
-        if selected_record is None:
-            return
-
-        if selected_action == VersionBrowserWidget.ACTION_OPEN:
-            backup_path = selected_record.backup_path
-            if backup_path is None:
-                MessageDialog(
-                    self._main_window,
-                    "The selected version has no backup file path.",
-                    "Open Version Failed",
-                ).exec_()
-                return
-            if not backup_path.exists() or not backup_path.is_file():
-                MessageDialog(
-                    self._main_window,
-                    f"Backup file is missing on disk:\n{backup_path}",
-                    "Open Version Failed",
-                ).exec_()
-                return
-            if not self._check_unsaved_changes():
-                return
-
-            load_warning: str | None = None
-            try:
-                load_warning = self._load_hip_file(backup_path)
-            except Exception as exc:
-                log.exception("Failed to load Houdini set version: %s", backup_path)
-                MessageDialog(
-                    self._main_window,
-                    (
-                        "Failed to open selected version:\n"
-                        f"{self._describe_exception(exc, fallback='Could not load the HIP file')}"
-                    ),
-                    "Open Version Failed",
-                ).exec_()
-                return
-
-            try:
-                self._post_open_file(environment)
-            except Exception as exc:
-                log.exception(
-                    "Loaded Houdini set version but post-open setup failed: %s",
-                    backup_path,
-                )
-                MessageDialog(
-                    self._main_window,
-                    (
-                        "The selected version loaded, but set setup could not finish:\n"
-                        f"{self._describe_exception(exc, fallback='Set post-open setup failed')}"
-                    ),
-                    "Open Version Failed",
-                ).exec_()
-                return
-
-            if load_warning:
-                self._show_hip_load_warning(
-                    path=backup_path,
-                    warning=load_warning,
-                    title="Version Opened With Warnings",
-                )
-            return
-
-        if selected_action == VersionBrowserWidget.ACTION_PROMOTE:
-            source_backup = selected_record.backup_path
-            if source_backup is None or not source_backup.exists():
-                MessageDialog(
-                    self._main_window,
-                    "Cannot create a new version from this entry because the backup file is missing.",
-                    "Create Version Failed",
-                ).exec_()
-                return
-
-            promote_dialog = PromoteVersionDialog(self._main_window, selected_record)
-            if not promote_dialog.exec_():
-                return
-            try:
-                promoted_record = promote_version_record(
-                    selected_record,
-                    set_stream,
-                    title=promote_dialog.get_title(),
-                    note=promote_dialog.get_note(),
-                )
-            except Exception as exc:
-                log.exception("Failed to create a new Houdini set version.")
-                MessageDialog(
-                    self._main_window,
-                    f"Failed to create new version:\n{exc}",
-                    "Create Version Failed",
-                ).exec_()
-                return
-
-            MessageDialog(
-                self._main_window,
-                (
-                    f"Created new version {version_label(promoted_record.version)} "
-                    f'"{promoted_record.title or "(untitled)"}" from the selected backup.\n'
-                    "Open it from Version History to continue working from it."
-                ),
-                "Version Created",
-            ).exec_()
 
     def save_version(self) -> None:
         hip_path = self._ensure_hip_saved()
         if hip_path is None:
             return
 
-        resolved = self._resolve_current_set_stream(hip_path)
+        resolved = self._resolve_current_stream(hip_path)
         if resolved is None:
             MessageDialog(
                 self._main_window,
@@ -326,32 +193,5 @@ class HEnvFileManager(HFileManager):
             ).exec_()
             return
 
-        _environment, set_stream = resolved
-        dialog = SaveVersionDialog(self._main_window)
-        if not dialog.exec_():
-            return
-
-        try:
-            version_record = save_version_record(
-                hip_path,
-                set_stream,
-                title=dialog.get_title(),
-                note=dialog.get_note(),
-            )
-        except Exception as exc:
-            log.exception("Failed to save Houdini set version.")
-            MessageDialog(
-                self._main_window,
-                f"Failed to save version:\n{exc}",
-                "Save Version Failed",
-            ).exec_()
-            return
-
-        MessageDialog(
-            self._main_window,
-            (
-                f"Saved {version_label(version_record.version)} "
-                f'"{version_record.title or "(untitled)"}".'
-            ),
-            "Version Saved",
-        ).exec_()
+        stream, _, _ = resolved
+        self._do_save_version(hip_path, stream)
