+++ pipeline/pipe/h/hipfile/filemanager.py
@@ -9,8 +9,17 @@ from env_sg import DB_Config
 import pipe.h
 from pipe.db import DB
 from pipe.glui.dialogs import MessageDialog
+from pipe.glui.save_version_dialog import PromoteVersionDialog, SaveVersionDialog
+from pipe.glui.version_browser import VersionBrowserWidget
 from pipe.struct.db import SGEntity
 from pipe.util import FileManager
+from pipe.versioning import (
+    VersionStreamSpec,
+    list_version_records,
+    promote_version as _promote_version,
+    save_version as _save_version,
+    version_label,
+)
 
 log = logging.getLogger(__name__)
 
@@ -27,6 +36,33 @@ class HFileManager(FileManager):
         super().__init__(
             conn, entity_type, window, versioning=versioning, version_glob=version_glob
         )
+
+    # ------------------------------------------------------------------
+    # Subclass contract
+    # ------------------------------------------------------------------
+
+    def _resolve_current_stream(
+        self, hip_path: Path
+    ) -> tuple[VersionStreamSpec, str, SGEntity] | None:
+        """Return (stream, owner_label, entity) for the current HIP, or None.
+
+        Subclasses must override this to resolve the versioning stream that
+        corresponds to the open HIP file.  ``owner_label`` is displayed in the
+        version browser header.  ``entity`` is passed to ``_post_open_file``
+        after opening a backup version.
+        """
+        raise NotImplementedError
+
+    def _entity_label(self) -> str:
+        """Human-readable noun for the entity kind managed by this class.
+
+        Used in dialog messages, e.g. ``"asset"``, ``"set"``, ``"shot"``.
+        """
+        return "file"
+
+    # ------------------------------------------------------------------
+    # Shared HIP helpers
+    # ------------------------------------------------------------------
 
     @staticmethod
     def _check_unsaved_changes() -> bool:
@@ -146,3 +182,179 @@ class HFileManager(FileManager):
             return None
 
         return hip_path
+
+    # ------------------------------------------------------------------
+    # Shared version browser and save
+    # ------------------------------------------------------------------
+
+    def open_version_browser(self) -> None:
+        kind = self._entity_label()
+        hip_path = self._current_hip_path()
+        if hip_path is None:
+            MessageDialog(
+                self._main_window,
+                f"No valid {kind} HIP is open. Use Open {kind.title()} first.",
+                "Version History",
+            ).exec_()
+            return
+
+        resolved = self._resolve_current_stream(hip_path)
+        if resolved is None:
+            MessageDialog(
+                self._main_window,
+                f"Could not resolve the current HIP to a valid {kind}. "
+                f"Use Open {kind.title()} first.",
+                "Version History",
+            ).exec_()
+            return
+
+        stream, owner_label, entity = resolved
+        records = list_version_records(stream)
+        if not records:
+            MessageDialog(
+                self._main_window,
+                f"No version history was found for this {kind}.",
+                "No Versions",
+            ).exec_()
+            return
+
+        browser = VersionBrowserWidget(
+            self._main_window,
+            records,
+            owner_label=owner_label,
+            stream_label=stream.label,
+        )
+        if not browser.exec_():
+            return
+
+        selected_record = browser.get_selected_record()
+        selected_action = browser.get_selected_action()
+        if selected_record is None:
+            return
+
+        if selected_action == VersionBrowserWidget.ACTION_OPEN:
+            backup_path = selected_record.backup_path
+            if backup_path is None:
+                MessageDialog(
+                    self._main_window,
+                    "The selected version has no backup file path.",
+                    "Open Version Failed",
+                ).exec_()
+                return
+            if not backup_path.exists() or not backup_path.is_file():
+                MessageDialog(
+                    self._main_window,
+                    f"Backup file is missing on disk:\n{backup_path}",
+                    "Open Version Failed",
+                ).exec_()
+                return
+            if not self._check_unsaved_changes():
+                return
+
+            load_warning: str | None = None
+            try:
+                load_warning = self._load_hip_file(backup_path)
+            except Exception as exc:
+                log.exception("Failed to open HIP version: %s", backup_path)
+                MessageDialog(
+                    self._main_window,
+                    (
+                        "Failed to open selected version:\n"
+                        f"{self._describe_exception(exc, fallback='Could not load the HIP file')}"
+                    ),
+                    "Open Version Failed",
+                ).exec_()
+                return
+
+            try:
+                self._post_open_file(entity)
+            except Exception as exc:
+                log.exception(
+                    "Loaded HIP version but post-open setup failed: %s", backup_path
+                )
+                MessageDialog(
+                    self._main_window,
+                    (
+                        f"The selected version loaded, but {kind} setup could not finish:\n"
+                        f"{self._describe_exception(exc, fallback=f'{kind.title()} post-open setup failed')}"
+                    ),
+                    "Open Version Failed",
+                ).exec_()
+                return
+
+            if load_warning:
+                self._show_hip_load_warning(
+                    path=backup_path,
+                    warning=load_warning,
+                    title="Version Opened With Warnings",
+                )
+            return
+
+        if selected_action == VersionBrowserWidget.ACTION_PROMOTE:
+            source_backup = selected_record.backup_path
+            if source_backup is None or not source_backup.exists():
+                MessageDialog(
+                    self._main_window,
+                    "Cannot create a new version from this entry because the backup file is missing.",
+                    "Create Version Failed",
+                ).exec_()
+                return
+
+            promote_dialog = PromoteVersionDialog(self._main_window, selected_record)
+            if not promote_dialog.exec_():
+                return
+            try:
+                promoted_record = _promote_version(
+                    selected_record,
+                    stream,
+                    title=promote_dialog.get_title(),
+                    note=promote_dialog.get_note(),
+                )
+            except Exception as exc:
+                log.exception("Failed to promote version.")
+                MessageDialog(
+                    self._main_window,
+                    f"Failed to create new version:\n{exc}",
+                    "Create Version Failed",
+                ).exec_()
+                return
+
+            MessageDialog(
+                self._main_window,
+                (
+                    f"Created new version {version_label(promoted_record.version)} "
+                    f'"{promoted_record.title or "(untitled)"}" from the selected backup.\n'
+                    "Open it from Version History to continue working from it."
+                ),
+                "Version Created",
+            ).exec_()
+
+    def _do_save_version(self, hip_path: Path, stream: VersionStreamSpec) -> None:
+        """Prompt for a version title and write a backup of *hip_path*."""
+        dialog = SaveVersionDialog(self._main_window)
+        if not dialog.exec_():
+            return
+        try:
+            version_record = _save_version(
+                hip_path,
+                stream,
+                title=dialog.get_title(),
+                note=dialog.get_note(),
+            )
+        except Exception as exc:
+            log.exception("Failed to save version.")
+            MessageDialog(
+                self._main_window,
+                f"Failed to save version:\n{exc}",
+                "Save Version Failed",
+            ).exec_()
+            return
+
+        MessageDialog(
+            self._main_window,
+            (
+                f"Saved {version_label(version_record.version)} "
+                f'"{version_record.title or "(untitled)"}".'
+            ),
+            "Version Saved",
+        ).exec_()
