+++ pipeline/pipe/h/hipfile/shot.py
@@ -10,21 +10,13 @@ import hou
 
 import pipe.h
 from pipe.glui.dialogs import FilteredListDialog, MessageDialog
-from pipe.glui.save_version_dialog import PromoteVersionDialog, SaveVersionDialog
-from pipe.glui.version_browser import VersionBrowserWidget
 from pipe.shot.version_adapter import (
     houdini_department_stream,
     path_matches_stream,
     shot_owner_for,
 )
 from pipe.struct.db import EnvironmentStub, SGEntity, Shot, validate_shot_code_token
-from pipe.versioning import (
-    VersionStreamSpec,
-    list_version_records,
-    promote_version,
-    save_version,
-    version_label,
-)
+from pipe.versioning import VersionStreamSpec
 
 from .filemanager import HFileManager
 
@@ -33,6 +25,9 @@ log = logging.getLogger(__name__)
 
 class HShotFileManager(HFileManager):
     _department: str | None
+
+    def _entity_label(self) -> str:
+        return "shot"
 
     class DEPARTMENT(str, Enum):
         CFX = "cfx"
@@ -178,153 +173,21 @@ class HShotFileManager(HFileManager):
             return None
         return shot, department, stream
 
-    def open_version_browser(self) -> None:
-        hip_path = self._current_hip_path()
-        if hip_path is None:
-            MessageDialog(
-                self._main_window,
-                "No valid shot HIP is open. Use Open Shot first.",
-                "Version History",
-            ).exec_()
-            return
-
+    def _resolve_current_stream(
+        self, hip_path: Path
+    ) -> tuple[VersionStreamSpec, str, SGEntity] | None:
         resolved = self._resolve_current_shot_stream(hip_path)
         if resolved is None:
-            MessageDialog(
-                self._main_window,
-                "Could not resolve the current HIP to a valid shot department file. Use Open Shot first.",
-                "Version History",
-            ).exec_()
-            return
-
-        shot, department, shot_stream = resolved
-        records = list_version_records(shot_stream)
-        if not records:
-            MessageDialog(
-                self._main_window,
-                "No version history was found for this shot department.",
-                "No Versions",
-            ).exec_()
-            return
-
-        browser = VersionBrowserWidget(
-            self._main_window,
-            records,
-            owner_label=f"{shot.code} ({department})",
-            stream_label=shot_stream.label,
-        )
-        if not browser.exec_():
-            return
-
-        selected_record = browser.get_selected_record()
-        selected_action = browser.get_selected_action()
-        if selected_record is None:
-            return
-
-        if selected_action == VersionBrowserWidget.ACTION_OPEN:
-            backup_path = selected_record.backup_path
-            if backup_path is None:
-                MessageDialog(
-                    self._main_window,
-                    "The selected version has no backup file path.",
-                    "Open Version Failed",
-                ).exec_()
-                return
-            if not backup_path.exists() or not backup_path.is_file():
-                MessageDialog(
-                    self._main_window,
-                    f"Backup file is missing on disk:\n{backup_path}",
-                    "Open Version Failed",
-                ).exec_()
-                return
-            if not self._check_unsaved_changes():
-                return
-
-            load_warning: str | None = None
-            try:
-                load_warning = self._load_hip_file(backup_path)
-            except Exception as exc:
-                log.exception("Failed to load Houdini shot version: %s", backup_path)
-                MessageDialog(
-                    self._main_window,
-                    (
-                        "Failed to open selected version:\n"
-                        f"{self._describe_exception(exc, fallback='Could not load the HIP file')}"
-                    ),
-                    "Open Version Failed",
-                ).exec_()
-                return
-
-            try:
-                self._post_open_file(shot)
-            except Exception as exc:
-                log.exception(
-                    "Loaded Houdini shot version but post-open setup failed: %s",
-                    backup_path,
-                )
-                MessageDialog(
-                    self._main_window,
-                    (
-                        "The selected version loaded, but shot setup could not finish:\n"
-                        f"{self._describe_exception(exc, fallback='Shot post-open setup failed')}"
-                    ),
-                    "Open Version Failed",
-                ).exec_()
-                return
-
-            if load_warning:
-                self._show_hip_load_warning(
-                    path=backup_path,
-                    warning=load_warning,
-                    title="Version Opened With Warnings",
-                )
-            return
-
-        if selected_action == VersionBrowserWidget.ACTION_PROMOTE:
-            source_backup = selected_record.backup_path
-            if source_backup is None or not source_backup.exists():
-                MessageDialog(
-                    self._main_window,
-                    "Cannot create a new version from this entry because the backup file is missing.",
-                    "Create Version Failed",
-                ).exec_()
-                return
-
-            promote_dialog = PromoteVersionDialog(self._main_window, selected_record)
-            if not promote_dialog.exec_():
-                return
-            try:
-                promoted_record = promote_version(
-                    selected_record,
-                    shot_stream,
-                    title=promote_dialog.get_title(),
-                    note=promote_dialog.get_note(),
-                )
-            except Exception as exc:
-                log.exception("Failed to create a new Houdini shot version.")
-                MessageDialog(
-                    self._main_window,
-                    f"Failed to create new version:\n{exc}",
-                    "Create Version Failed",
-                ).exec_()
-                return
-
-            MessageDialog(
-                self._main_window,
-                (
-                    f'Created new version {version_label(promoted_record.version)} '
-                    f'"{promoted_record.title or "(untitled)"}" from the selected backup.\n'
-                    "Open it from Version History to continue working from it."
-                ),
-                "Version Created",
-            ).exec_()
+            return None
+        shot, department, stream = resolved
+        return stream, f"{shot.code} ({department})", shot
 
     def save_version(self) -> None:
         hip_path = self._ensure_hip_saved()
         if hip_path is None:
             return
 
-        resolved = self._resolve_current_shot_stream(hip_path)
+        resolved = self._resolve_current_stream(hip_path)
         if resolved is None:
             MessageDialog(
                 self._main_window,
@@ -333,35 +196,8 @@ class HShotFileManager(HFileManager):
             ).exec_()
             return
 
-        _shot, _department, shot_stream = resolved
-        dialog = SaveVersionDialog(self._main_window)
-        if not dialog.exec_():
-            return
-
-        try:
-            version_record = save_version(
-                hip_path,
-                shot_stream,
-                title=dialog.get_title(),
-                note=dialog.get_note(),
-            )
-        except Exception as exc:
-            log.exception("Failed to save Houdini shot version.")
-            MessageDialog(
-                self._main_window,
-                f"Failed to save version:\n{exc}",
-                "Save Version Failed",
-            ).exec_()
-            return
-
-        MessageDialog(
-            self._main_window,
-            (
-                f'Saved {version_label(version_record.version)} '
-                f'"{version_record.title or "(untitled)"}".'
-            ),
-            "Version Saved",
-        ).exec_()
+        stream, _, _ = resolved
+        self._do_save_version(hip_path, stream)
 
     def _shot_setup_payload(self, *, shot: Shot, path: Path) -> dict[str, object]:
         return {
